Move system property into server
