+++ controller/src/main/resources/org/jboss/as/controller/descriptions/common/LocalDescriptions.properties
@@ -249,12 +249,6 @@ local-destination-outbound-socket-binding.fixed-source-port=Whether the port val
 # Subsystem
 subsystem.describe=Outputs the subsystem as a list of operations that can be executed to create the original model
 
-# System Properties
-system-property.add=Adds a system property or updates an existing one.
-system-property.remove=Removes a system property.
-system-property.value=The value of the system property.
-system-property.boottime=If true the system property is passed on the command-line to the started server jvm. If false, it will be pushed to the server as part of the startup sequence.
-
 # Deployments
 deployment=A deployment represents anything that can be deployed (e.g. an application such as EJB-JAR, WAR, EAR, any kind of standard archive such as RAR or JBoss-specific deployment) into a server.
 deployment.name=Unique identifier of the deployment. Must be unique across all deployments.
