+++ host-controller/src/main/java/org/jboss/as/host/controller/HostControllerEnvironment.java
@@ -703,14 +703,14 @@ public class HostControllerEnvironment extends ProcessEnvironment {
     }
 
     @Override
-    protected boolean isRuntimeSystemPropertyUpdateAllowed(String propertyName, String propertyValue, boolean bootTime) {
+    public boolean isRuntimeSystemPropertyUpdateAllowed(String propertyName, String propertyValue, boolean bootTime) {
         // Currently any system-property in host.xml should not be applied to the HC runtime. This method
         // should not be invoked.
         throw MESSAGES.hostControllerSystemPropertyUpdateNotSupported();
     }
 
     @Override
-    protected void systemPropertyUpdated(String propertyName, String propertyValue) {
+    public void systemPropertyUpdated(String propertyName, String propertyValue) {
         // no-op
     }
