+++ host-controller/src/main/java/org/jboss/as/host/controller/descriptions/HostRootDescription.java
@@ -89,6 +89,7 @@ import org.jboss.as.host.controller.operations.HostShutdownHandler;
 import org.jboss.as.host.controller.operations.LocalDomainControllerAddHandler;
 import org.jboss.as.host.controller.operations.RemoteDomainControllerAddHandler;
 import org.jboss.as.host.controller.operations.RemoteDomainControllerRemoveHandler;
+import org.jboss.as.server.controller.descriptions.SystemPropertyDescriptions;
 import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
 
@@ -340,7 +341,7 @@ public class HostRootDescription {
 
     public static ModelNode getSystemPropertiesDescription(Locale locale) {
         final ResourceBundle bundle = getResourceBundle(locale);
-        return CommonDescriptions.getSystemPropertyDescription(locale, bundle.getString("host.system-property"), true);
+        return SystemPropertyDescriptions.getSystemPropertyDescription(locale, bundle.getString("host.system-property"), true);
     }
 
     public static ModelNode getLocalDomainControllerAdd(Locale locale) {
