+++ server/src/main/resources/org/jboss/as/server/controller/descriptions/LocalDescriptions.properties
@@ -96,6 +96,12 @@ vault.vault-options=Security Vault options.
 vault.add=Add the Security Vault.
 vault.remove=Remove the Security Vault.
 
+# System Properties
+system-property.add=Adds a system property or updates an existing one.
+system-property.remove=Removes a system property.
+system-property.value=The value of the system property.
+system-property.boottime=If true the system property is passed on the command-line to the started server jvm. If false, it will be pushed to the server as part of the startup sequence.
+
 # Operations
 composite=An operation that executes a list of other operations as a unit.
 composite.steps=A list, where each item in the list has the same structure as a simple operation request.
